+++ Mila/Tests/Dnn/Tensors/TensorOps/Fill.Cpu.cpp
@@ -166,6 +166,6 @@ namespace Dnn::Tensors::TensorOps::Tests
         EXPECT_EQ( scalar.item(), 999 );
 
         // Verify operator[] throws for scalars
-        EXPECT_THROW( scalar[std::vector<size_t>{}], std::runtime_error );
+        EXPECT_THROW( scalar[std::vector<int64_t>{}], std::runtime_error );
     }
 }
