+++ Mila/Tests/Dnn/Tensors/TensorOps/Math.Cuda.cpp
@@ -53,7 +53,7 @@ namespace Dnn::Tensors::TensorOps::Tests
     template<TensorDataType TDataType>
     Tensor<TDataType, CudaDeviceMemoryResource> makeCudaTensor(
         const std::string& device_name,
-        const std::vector<size_t>& shape,
+        const std::vector<int64_t>& shape,
         const std::vector<typename TensorHostTypeMap<TDataType>::host_type>& values = {} )
     {
         Tensor<TDataType, CudaDeviceMemoryResource> tensor( device_name, shape );
